Rename timeseries module
