+++ ptsa/data/timeseries.py
