+++ tests/test_filters.py
@@ -5,7 +5,7 @@ import pytest
 
 import numpy as np
 from numpy.testing import assert_array_equal, assert_array_almost_equal
-from ptsa.data import TimeSeriesX
+from ptsa.data import timeseries
 from ptsa.data.readers import BaseEventReader
 from ptsa.data.filters.MorletWaveletFilter import MorletWaveletFilter
 from ptsa.data.readers.tal import TalReader
@@ -156,7 +156,7 @@ class TestFiltersExecute(unittest.TestCase):
     def setUp(self):
         times = np.linspace(0,1,1000)
         ts = np.sin(8*times) + np.sin(16*times) + np.sin(32*times)
-        self.time_series = TimeSeriesX.TimeSeriesX(data=ts,dims=('time'),coords = {'time':times,'samplerate':1000})
+        self.time_series = timeseries.TimeSeriesX(data=ts, dims=('time'), coords = {'time':times, 'samplerate':1000})
 
 
     def test_ButterworthFilter(self):
